Transpile third-party JSX
This is required when importing JSX from third-party dependencies (e.g.
another components library).

Babel reasonably ignores `node_modules` by default, we now whitelist
everything by blanking the `ignore` option. Unfortunately, due to the
global nature of @babel/register, there isn't really a good way to make
this configurable in order to only whitelist individual dependencies.

I'm not sure the effect (if any) that this will have on the performance
of the adapter, but I'm not too worried about it as transpilation should
only occur when a module is first loaded.